итерация 3

+++ CPLUS-13.7.3/tree.cpp
@@ -65,12 +65,12 @@ void Tree::print()
 void Tree::printInner(Node *current)
 {
    std::cout << "N" << current->number << ": ";
-   for (size_t i = 0; i < current->childCount; i++)
+   for (int i = 0; i < current->childCount; i++)
    {
       std::cout << current->children[i]->number << ", ";
    }
    std::cout << std::endl;
-   for (size_t i = 0; i < current->childCount; i++)
+   for (int i = 0; i < current->childCount; i++)
    {
       printInner(current->children[i]);
    }
